my_requirements.txt
# python=3.6
jupyter
numpy
# pytorch=1.3
torchvision
pandas

cython
pandas
tqdm
spacy
redis
pyzmq
ujson
# torch-scatter==1.3.2	# merged into SpeakQL
msgpack
h5py
fairseq==0.8.0

# transformers
--editable=git+https://github.com/huggingface/transformers.git@372a5c1ceec49b52c503707e9657bfaae7c236a0#egg=pytorch_pretrained_bert
tokenizers==0.8.1   # Otherwise incompatible with Mac 10.13; shouldn't need this on server

--editable .